refactor(backend): split farm/cultures/views.py into a package
Split the 888-line cultures views module into one module per sub-area
inside farm/cultures/views/: suppliers (SupplierViewSet,
CultureSupplierDataViewSet), cultures (CultureViewSet,
CultureUndeleteView), public (PublicCultureViewSet), seed_packages
(SeedPackageViewSet) and seed_demand (SeedDemandListView).

All bodies are moved verbatim; the package __init__ re-exports the view
names so the farm.cultures.views import path used by farm/urls.py keeps
working unchanged. The module-level logger, which was never used, is
dropped. Ruff statistics are identical to main; the URL contract
(pattern + name) is unchanged, only the view module paths differ.
