+++ backend/farm/cultures/views/__init__.py
@@ -1,0 +1,21 @@
+"""API endpoints for the cultures domain (cultures, suppliers, seeds, public library).
+
+Split into one module per sub-area; this package re-exports the view names so
+`farm.cultures.views` keeps working as an import path (used by farm/urls.py).
+"""
+
+from .cultures import CultureUndeleteView, CultureViewSet
+from .public import PublicCultureViewSet
+from .seed_demand import SeedDemandListView
+from .seed_packages import SeedPackageViewSet
+from .suppliers import CultureSupplierDataViewSet, SupplierViewSet
+
+__all__ = [
+    'CultureSupplierDataViewSet',
+    'CultureUndeleteView',
+    'CultureViewSet',
+    'PublicCultureViewSet',
+    'SeedDemandListView',
+    'SeedPackageViewSet',
+    'SupplierViewSet',
+]
