+++ backend/farm/cultures/views/seed_packages.py
@@ -1,0 +1,42 @@
+"""API endpoints for seed packages."""
+
+
+from rest_framework import viewsets
+
+from farm.common.mixins import ProjectRevisionMixin, ProjectScopedMixin
+from farm.history import (
+    _entity_display_name,
+    _serialize_instance,
+)
+from farm.models import (
+    EntityRevision,
+    SeedPackage,
+)
+
+from ..serializers import (
+    SeedPackageSerializer,
+)
+
+
+class SeedPackageViewSet(ProjectScopedMixin, ProjectRevisionMixin, viewsets.ModelViewSet):
+    queryset = SeedPackage.objects.select_related('culture').all().order_by('size_unit', 'size_value')
+    serializer_class = SeedPackageSerializer
+
+    def perform_create(self, serializer):
+        instance = serializer.save(project=self.request.active_project)
+        self.record_revision(instance, EntityRevision.ACTION_CREATED)
+
+    def perform_update(self, serializer):
+        previous_snapshot = _serialize_instance(serializer.instance)
+        instance = serializer.save()
+        self.record_revision(instance, EntityRevision.ACTION_UPDATED, previous_snapshot=previous_snapshot)
+
+    def perform_destroy(self, instance):
+        package_id = instance.pk
+        snapshot = _serialize_instance(instance)
+        display_name = _entity_display_name(instance)
+        instance.delete()
+        self.record_revision(
+            instance, EntityRevision.ACTION_DELETED,
+            object_id=package_id, snapshot=snapshot, display_name=display_name, changed_fields=[],
+        )
