+++ backend/farm/cultures/views/suppliers.py
@@ -1,0 +1,214 @@
+"""API endpoints for suppliers and per-culture supplier data."""
+
+
+from django.db import IntegrityError, transaction
+from rest_framework import status, viewsets
+from rest_framework.decorators import action
+from rest_framework.exceptions import ValidationError as DRFValidationError
+from rest_framework.request import Request
+from rest_framework.response import Response
+
+from farm.common.mixins import ProjectRevisionMixin, ProjectScopedMixin
+from farm.history import (
+    _entity_display_name,
+    _serialize_instance,
+)
+from farm.models import (
+    CultureSupplierData,
+    EntityRevision,
+    Supplier,
+)
+from farm.services.suppliers import (
+    DuplicateSupplierNameError,
+    SupplierPayloadError,
+    SupplierRestoreConflictError,
+    SupplierRestoreFailedError,
+    build_delete_undo_payload,
+    build_delete_usage,
+    create_supplier,
+    normalize_new_supplier_payload,
+    restore_unlinked_supplier,
+    unlink_supplier_references,
+)
+
+from ..serializers import (
+    CultureSupplierDataSerializer,
+    SupplierSerializer,
+)
+
+
+class SupplierViewSet(ProjectScopedMixin, ProjectRevisionMixin, viewsets.ModelViewSet):
+    """ViewSet for Supplier model providing CRUD operations.
+    
+    Provides list, create, retrieve, update, and delete operations
+    for seed suppliers. Supports filtering by name via query parameter.
+    POST endpoint rejects duplicate names within the active project.
+    
+    Attributes:
+        queryset: All Supplier objects ordered by name
+        serializer_class: SupplierSerializer for serialization
+    """
+    queryset = Supplier.objects.all()
+    serializer_class = SupplierSerializer
+
+    def perform_update(self, serializer):
+        previous_snapshot = _serialize_instance(serializer.instance)
+        try:
+            instance = serializer.save()
+        except IntegrityError as exc:
+            raise DRFValidationError({'name': ['Ein Lieferant mit diesem Namen existiert bereits.']}) from exc
+        self.record_revision(instance, EntityRevision.ACTION_UPDATED, previous_snapshot=previous_snapshot)
+
+    @action(detail=True, methods=['get'], url_path='delete-usage')
+    def delete_usage(self, request: Request, pk: int | None = None) -> Response:
+        supplier = self.get_object()
+        return Response(build_delete_usage(supplier))
+
+    def destroy(self, request: Request, *args: object, **kwargs: object) -> Response:
+        instance = self.get_object()
+        usage = build_delete_usage(instance)
+        if not usage['can_delete']:
+            return Response(
+                {
+                    'detail': 'Supplier is still used and cannot be deleted.',
+                    'usage': usage,
+                },
+                status=status.HTTP_409_CONFLICT,
+            )
+        self.perform_destroy(instance)
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+    @action(detail=True, methods=['post'], url_path='unlink-and-delete')
+    def unlink_and_delete(self, request: Request, pk: int | None = None) -> Response:
+        supplier = self.get_object()
+        usage = build_delete_usage(supplier)
+        undo_payload = build_delete_undo_payload(supplier)
+
+        with transaction.atomic():
+            unlink_supplier_references(supplier)
+            supplier_id = supplier.pk
+            supplier_snapshot = _serialize_instance(supplier)
+            supplier_name = supplier.name
+            supplier.delete()
+            self.record_revision(
+                supplier, EntityRevision.ACTION_DELETED,
+                object_id=supplier_id, snapshot=supplier_snapshot, display_name=supplier_name, changed_fields=[],
+            )
+
+        return Response({
+            'affected_culture_count': usage['total_culture_count'],
+            'undo_payload': undo_payload,
+        })
+
+    @action(detail=False, methods=['post'], url_path='restore-unlinked-delete')
+    def restore_unlinked_delete(self, request: Request) -> Response:
+        payload = request.data if isinstance(request.data, dict) else {}
+        try:
+            result = restore_unlinked_supplier(
+                project=request.active_project,
+                payload=payload,
+                record_restore=lambda supplier: self.record_revision(supplier, EntityRevision.ACTION_RESTORED),
+            )
+        except SupplierPayloadError as exc:
+            raise DRFValidationError(exc.errors) from exc
+        except SupplierRestoreConflictError:
+            return Response(
+                {'detail': 'Supplier cannot be restored because the id is already in use.'},
+                status=status.HTTP_409_CONFLICT,
+            )
+        except SupplierRestoreFailedError as exc:
+            raise DRFValidationError({'detail': ['Supplier could not be restored.']}) from exc
+
+        serializer = self.get_serializer(result.supplier)
+        return Response({
+            'supplier': serializer.data,
+            'restored_culture_count': result.restored_culture_count,
+            'restored_supplier_data_count': result.restored_supplier_data_count,
+        })
+
+    def perform_destroy(self, instance: Supplier) -> None:
+        instance_id = instance.pk
+        snapshot = _serialize_instance(instance)
+        name = instance.name
+        instance.delete()
+        self.record_revision(
+            instance, EntityRevision.ACTION_DELETED,
+            object_id=instance_id, snapshot=snapshot, display_name=name, changed_fields=[],
+        )
+
+    
+    def get_queryset(self):
+        """Filter suppliers by name if query parameter is provided.
+        
+        :return: Filtered queryset based on query parameters
+        """
+        queryset = super().get_queryset()
+        query = self.request.query_params.get('q', None)
+
+        if query:
+            # Case-insensitive search in name
+            queryset = queryset.filter(name__icontains=query)
+
+        queryset = queryset.order_by('name')
+
+        # Limit only list responses for autocomplete-like usage.
+        # Detail/update/delete must be able to resolve any existing supplier by PK.
+        if getattr(self, 'action', None) == 'list':
+            return queryset[:20]
+
+        return queryset
+    
+    def create(self, request, *args, **kwargs):
+        """Create a supplier with project-scoped duplicate-name validation.
+
+        :param request: HTTP request containing supplier data
+        :return: Response with supplier data and created flag
+        """
+        try:
+            fields = normalize_new_supplier_payload(
+                name=request.data.get('name'),
+                homepage_url=request.data.get('homepage_url'),
+                allowed_domains=request.data.get('allowed_domains', []),
+            )
+            supplier = create_supplier(project=request.active_project, **fields)
+        except SupplierPayloadError as exc:
+            return Response(exc.errors, status=status.HTTP_400_BAD_REQUEST)
+        except DuplicateSupplierNameError as exc:
+            raise DRFValidationError({'name': ['Ein Lieferant mit diesem Namen existiert bereits.']}) from exc
+
+        serializer = self.get_serializer(supplier)
+        data = serializer.data
+        data['created'] = True
+
+        self.record_revision(supplier, EntityRevision.ACTION_CREATED)
+        return Response(
+            data,
+            status=status.HTTP_201_CREATED
+        )
+
+
+class CultureSupplierDataViewSet(ProjectScopedMixin, ProjectRevisionMixin, viewsets.ModelViewSet):
+    queryset = CultureSupplierData.objects.select_related('culture', 'supplier')
+    serializer_class = CultureSupplierDataSerializer
+
+    def get_queryset(self):
+        return self.queryset.filter(project=self.request.active_project)
+
+    def perform_create(self, serializer):
+        instance = serializer.save(project=self.request.active_project)
+        self.record_revision(instance, EntityRevision.ACTION_CREATED)
+
+    def perform_update(self, serializer):
+        previous_snapshot = _serialize_instance(serializer.instance)
+        instance = serializer.save()
+        self.record_revision(instance, EntityRevision.ACTION_UPDATED, previous_snapshot=previous_snapshot)
+
+    def perform_destroy(self, instance):
+        instance_id = instance.pk
+        snapshot = _serialize_instance(instance)
+        display_name = _entity_display_name(instance)
+        instance.delete()
+        self.record_revision(
+            instance, EntityRevision.ACTION_DELETED,
+            object_id=instance_id, snapshot=snapshot, display_name=display_name, changed_fields=[],
+        )
